refactor: ♻️ decompose EntityList filters and search to smaller components
